refactor: move validator helper to its own file
