+++ src/aind_data_schema/core/procedures.py
@@ -30,7 +30,7 @@ from aind_data_schema.components.devices import FiberProbe, MyomatrixArray
 from aind_data_schema.components.identifiers import Person
 from aind_data_schema.components.reagent import Reagent
 from aind_data_schema.utils.merge import merge_notes
-from aind_data_schema.utils.compatibility_check import subject_specimen_id_compatibility
+from aind_data_schema.utils.validators import subject_specimen_id_compatibility
 
 
 class ImmunolabelClass(str, Enum):
