+++ src/aind_data_schema/utils/validators.py
@@ -1,0 +1,6 @@
+""" Validator utility functions """
+
+
+def subject_specimen_id_compatibility(subject_id: str, specimen_id: str) -> bool:
+    """Check whether a subject_id and specimen_id are compatible"""
+    return subject_id in specimen_id
